Script: Tapscript v2: add varcost for DUP operations
These are the same as other script versions, but we do need to account for the
cost of the pushes.

BIKESHED: We don't charge for *all* push_back operations
(e.g. OP_VERIFY, PUSH opcodes), so we don't simply replace push_back.
But we could change the BIP if that was simpler?

TODO: Tests
Signed-off-by: Rusty Russell <rusty@rustcorp.com.au>