script/interpreter: Tapscript v2: interpret numbers as varnums not CScriptNum (non-arithmetic opcodes).
The following non-arithmetic opcodes interpret numbers on the stack as CScriptNum:
1. OP_CHECKLOCKTIMEVERIFY.
2. OP_CHECKSEQUENCEVERIFY.
3. OP_VERIFY.
4. OP_ROLL.
5. OP_CHECKSIGADD.

In the first two cases, all numbers > 2^32-1 are equivalent, and we can easily represent them as CScriptNum (which is an int64_t internally).

OP_VERIFY uses CastToBool which unfortunately treats "negative zero" as false.

OP_ROLL is invalid if the value is greater or equal to the stack depth, so we can cap value conversion at that.

OP_CHECKSIGADD doesn't really care about the number, it just adds 1 and puts it back.